everything works, need to clean up more
