+++ src/components/home/index.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import { useStateValue } from '../../state'
 import Header from '../header'
-import Calories from '../calories/Calories'
-import GoalBar from '../goal-bar/GoalBar'
+import Calories from '../calories'
+import GoalBar from '../goal-bar'
 import Controls from '../controls'
-import Modify from '../modify/Modify'
-import LogOut from '../logout/LogOut'
+import Modify from '../modify'
+import LogOut from '../logout'
 
 const Home = () => {
     const [{ user }] = useStateValue()
@@ -27,3 +27,28 @@ const Home = () => {
 }
 
 export default Home
+
+
+//   detectKeyDown = e => {
+//     switch( e.key ) {
+//       case ']':
+//         this.toggleAddCalories()
+//         break
+//       case '[':
+//         this.toggleSubtractCalories()
+//         break
+//       case 'Escape':
+//         this.setState({
+//           toggle_add: false,
+//           toggle_subtract: false,
+//           toggle_logout: false,
+//           toggle_change_goal: false
+//         })
+//         break
+//       case 'g':
+//         this.toggleChangeGoal()
+//         break
+//       default:
+//         break
+//     }
+//   }
