+++ src/components/logout/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Footer from '../footer/Footer'
+import Footer from '../footer'
 import { useStateValue } from '../../state'
 import { auth } from '../../utils/firebase'
