+++ src/components/shared/progress-bar/_progress-bar.scss
@@ -3,6 +3,7 @@
   top: 0;
   left: 0;
   width: 0%;
+  max-width: 100%;
   height: 100%;
   z-index: 10;
   transition: all 0.4s ease;
