+++ src/components/shared/progress-bar/index.js
