+++ src/reducers/index.js
@@ -1,0 +1,28 @@
+export const rootReducer = (state, action) => {
+    console.log(action)
+    switch (action.type) {
+      case "user":
+        return {
+          ...state,
+          user: action.payload
+        }
+      case "toggleLoader":
+        return {
+          ...state,
+          isLoading: action.payload
+        }
+      case "toggleModal":
+        return {
+          ...state,
+          isModalActive: action.payload
+        }
+      case "modify":
+        return {
+          ...state,
+          modify: action.payload
+        }
+
+      default:
+        return state
+    }
+  }
