+++ src/state/index.js
@@ -2,6 +2,25 @@ import React, { createContext, useContext, useReducer } from 'react'
 
 export const StateContext = createContext()
 
+export const initialState = {
+  user: {
+    isAuthenticated: false,
+    uid: null,
+    username: null,
+    avatar: null,
+    totalCalories: 0,
+    dailyGoal: 0,
+    history: []
+  },
+  modify: {
+    add: false,
+    subtract: false,
+    edit: false
+  },
+  isModalActive: false,
+  isLoading: false
+}
+
 export const StateProvider = ({ reducer, initialState, children }) => (
   <StateContext.Provider value={useReducer(reducer, initialState)}>
     {children}
